Refactor to a transient event consumer
